backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
drf_spectacular
django-filter